site/backend/pyproject.toml
[project]
name = "chess-site-backend"
version = "0.1.0"
description = "Backend for Chess Engine Site"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-chess",
    "grpcio",
    "grpcio-tools",
    "protobuf"
]

